fix: custom implementation of the IntOrString type
Kubernetes generates its OpenAPI spec using v2, which suffers from some
limitations.

We then generate all our Go types using go-swagger, which currently
supports only OpenAPI v2.

Prior to this commit, the `IntOrString` type was generated by go-swagger
as an alias of `string`.

This is a problem, because trying to unmarshal something that is a
number results in a json error.

The proper fix would be to have kubernetes upstream generate OpenAPI v3 spec files
and then use a version of go-swagger that supports OpenAPI v3.

None of that is doable right now, hence with this commit we provide a
custom implementation of the `IntOrString` type that works.

Fixes https://github.com/kubewarden/k8s-objects-generator/issues/34

Signed-off-by: Flavio Castelli <fcastelli@suse.com>